Implement check script
Add client package. This package contains the interface definition
for a calendar client. Calendar clients encapsulate all of the interaction
with the calendar provider. The only provider implemented so far is
Google Calendar, using their own API client library.

Add calendar package. This holds the events retrieved from a calendar
and owns all of the logic for calculating whether an event is currently
happening, which in Concourse resource terms means it is a valid 'Version'
of the resource. A TimeKeeper interface was introduced so a fake timer
can be provided for testing purposes.

Add errors package. Very little error handling is implemented, as the
Concourse resource interface requires that any failure to satisfy the
`check`, `in`, or `out` procedures should result in the scripts exiting
with an error, so the same error handling function is used throughout.

Add models package. This contains the data structures relating to input
and output to the scripts.