Add `in` script
For a resource version to be used inside a task it must provide an
`in` script which takes a resource version on standard input, emits
the version being fetched, along with any useful metadata, and place
a file in the directory passed to it as argument `$1`.

The interface method CalendarClient.GetEvent() must provide both the
response and the file, as the metadata which is useful and the contents
of the file could vary between calendar providers.