Refactor how resource versions are identified
Previously a current version of a resource was defined by a struct
containing the event's start and end time. The problem was this does
not uniquely identify an event. Also, the calendar client was getting
all of the events in the calendar that haven't finished yet, when we
are really only interested in current events. The API can determine what
events are current with very little logic, therefore removal of the calendar
package and tests is justified.

Resource versions are now just idenfitied by their event ID, which is
unique for all events, even individual recurring events[1].

[1] https://godoc.org/google.golang.org/api/calendar/v3#Event